updated db refactor
